refactor: CarDict 디렉토리 구조 수정 및 import 수정
