+++ Frontend/src/components/car-dict/CarDictBox/CarDictBox.style.tsx
