+++ Frontend/src/components/car-dict/CarDictModal/CarDictModal.style.tsx
@@ -1,0 +1,3 @@
+import { styled } from 'styled-components';
+
+export const ModalContainer = styled.div``;
