+++ Frontend/src/components/car-dict/CarDictModal/CarDictModal.tsx
@@ -1,0 +1,9 @@
+import * as S from './CarDictModal.style';
+
+export default function CarDictModal() {
+  return (
+    <>
+      <S.ModalContainer></S.ModalContainer>
+    </>
+  );
+}
